+++ package.json
@@ -9,6 +9,7 @@
     "gh-pages": "^1.1.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.0.17",
     "react-stars": "^2.2.5",
     "react-test-renderer": "^16.2.0"
