+++ src/components/main/Picture.css
