+++ src/components/main/Picture.jsx
