+++ src/components/main/Quote.jsx
