+++ src/components/main/Stars.jsx
